+++ example_packages/hello_world/build.py.example
@@ -1,3 +1,10 @@
+"""
+To use this as the build script, add to package.py:
+
+    build_command = 'python {root}/build.py {install}'
+
+"""
+
 import os
 import os.path
 import shutil
@@ -44,3 +51,13 @@ def build(source_path, build_path, install_path, targets):
 
     if "install" in (targets or []):
         _install()
+
+
+if __name__ == '__main__':
+    import os, sys
+    build(
+        source_path=os.environ['REZ_BUILD_SOURCE_PATH'],
+        build_path=os.environ['REZ_BUILD_PATH'],
+        install_path=os.environ['REZ_BUILD_INSTALL_PATH'],
+        targets=sys.argv[1:]
+    )
