+++ src/rez/tests/data/builds/packages/anti/1.0.0/build.py
@@ -1,29 +1,32 @@
 from __future__ import print_function
 
-import shutil
-import os.path
-
+from build_util import build_directory_recurse, check_visible
 
 def build(source_path, build_path, install_path, targets):
 
-    def _copy(src, dest):
-        print("copying %s to %s..." % (src, dest))
-        if os.path.exists(dest):
-            shutil.rmtree(dest)
-        shutil.copytree(src, dest)
+    # normal requirement 'foo' should be visible
+    check_visible('anti', 'build_util')
 
-    # build
-    src = os.path.join(source_path, "build_util")
-    dest = os.path.join(build_path, "python", "build_util")
-    _copy(src, dest)
+    check_visible('anti', 'floob')
+    import floob
+    floob.hello()
 
-    if "install" not in (targets or []):
-        return
+    try:
+        import loco
+        raise Exception('loco should not be here')
+    except ImportError:
+        print('Intentionally raising an ImportError since loco should not be available')
+        pass
 
-    # install
-    src = os.path.join(build_path, "python")
-    dest = os.path.join(install_path, "python")
-    _copy(src, dest)
+
+if __name__ == '__main__':
+    import os, sys
+    build(
+        source_path=os.environ['REZ_BUILD_SOURCE_PATH'],
+        build_path=os.environ['REZ_BUILD_PATH'],
+        install_path=os.environ['REZ_BUILD_INSTALL_PATH'],
+        targets=sys.argv[1:]
+    )
 
 
 # Copyright 2013-2016 Allan Johns.
