+++ src/rez/tests/data/builds/packages/build_util/1/build.py
@@ -13,17 +13,27 @@ def build(source_path, build_path, install_path, targets):
         shutil.copytree(src, dest)
 
     # build
-    src = os.path.join(source_path, "data")
-    dest = os.path.join(build_path, "data")
+    src = os.path.join(source_path, "build_util")
+    dest = os.path.join(build_path, "python", "build_util")
     _copy(src, dest)
 
     if "install" not in (targets or []):
         return
 
     # install
-    src = os.path.join(build_path, "data")
-    dest = os.path.join(install_path, "data")
+    src = os.path.join(build_path, "python")
+    dest = os.path.join(install_path, "python")
     _copy(src, dest)
+
+
+if __name__ == '__main__':
+    import os, sys
+    build(
+        source_path=os.environ['REZ_BUILD_SOURCE_PATH'],
+        build_path=os.environ['REZ_BUILD_PATH'],
+        install_path=os.environ['REZ_BUILD_INSTALL_PATH'],
+        targets=sys.argv[1:]
+    )
 
 
 # Copyright 2013-2016 Allan Johns.
