+++ src/rez/tests/data/builds/packages/build_util/1/package.py
@@ -7,6 +7,8 @@ description = "build utilities"
 def commands():
     env.PYTHONPATH.append('{root}/python')
 
+build_command = 'python {root}/build.py {install}'
+
 
 # Copyright 2013-2016 Allan Johns.
 #
