+++ src/rez/tests/data/builds/packages/floob/package.py
@@ -13,6 +13,8 @@ private_build_requires = ["build_util"]
 def commands():
     env.PYTHONPATH.append('{root}/python')
 
+build_command = 'python {root}/build.py {install}'
+
 
 # Copyright 2013-2016 Allan Johns.
 #
