+++ src/rez/tests/data/builds/packages/foo/1.0.0/package.py
@@ -14,6 +14,8 @@ def pre_build_commands():
 def commands():
     env.PYTHONPATH.append('{root}/python')
 
+build_command = 'python {root}/build.py {install}'
+
 
 # Copyright 2013-2016 Allan Johns.
 #
