+++ src/rez/tests/data/builds/packages/loco/3/build.py
@@ -1,0 +1,2 @@
+
+# intentionally empty
