+++ src/rez/tests/data/builds/packages/whack/build.py
@@ -1,17 +1,18 @@
-from build_util import build_directory_recurse
-import os.path
+import os
+import sys
 
 
 def build(source_path, build_path, install_path, targets):
+    raise Exception("This build is deliberately broken")
 
-    if "install" not in (targets or []):
-        install_path = None
 
-    build_directory_recurse(src_dir="floob",
-                            dest_dir=os.path.join("python", "floob"),
-                            source_path=source_path,
-                            build_path=build_path,
-                            install_path=install_path)
+if __name__ == '__main__':
+    build(
+        source_path=os.environ['REZ_BUILD_SOURCE_PATH'],
+        build_path=os.environ['REZ_BUILD_PATH'],
+        install_path=os.environ['REZ_BUILD_INSTALL_PATH'],
+        targets=sys.argv[1:]
+    )
 
 
 # Copyright 2013-2016 Allan Johns.
