+++ src/rez/tests/data/release/package.yaml
@@ -4,3 +4,5 @@ authors:
 - joe.bloggs
 uuid: '9982b60993af4a4d89e8372472a49d02'
 description: 'foo type thing'
+
+build_command: "python {root}/build.py {install}"
