+++ src/rez/tests/data/release/variants/build.py
@@ -1,0 +1,2 @@
+
+# intentionally empty
