+++ src/rez/tests/data/release/variants/package.yaml
@@ -6,3 +6,5 @@ description: 'a package with two variants'
 variants:
 - ['spangle-1.0']
 - ['spangle-1.1']
+
+build_command: "python {root}/build.py {install}"
