--- src/rez/tests/data/release/variants/rezbuild.py
@@ -1,6 +1,0 @@
-import shutil
-import os.path
-
-
-def build(source_path, build_path, install_path, targets):
-    pass
