+++ src/rez/utils/_version.py
@@ -1,7 +1,7 @@
 
 
 # Update this value to version up Rez. Do not place anything else in this file.
-_rez_version = "2.69.6"
+_rez_version = "2.70.0"
 
 
 # Copyright 2013-2016 Allan Johns.
